Move BatchQueue to the aws app.
