+++ .env.dist
@@ -64,7 +64,7 @@ SECRET_KEY=
 # queue, then use this.
 # Requires S3 media storage, not local storage.
 # Also, you should set SPACER_JOB_HASH if using this.
-#SPACER_QUEUE_CHOICE=vision_backend.queues.BatchQueue
+#SPACER_QUEUE_CHOICE=aws.queues.BatchQueue
 
 # Normally we make huey run async tasks immediately after they're queued for
 # convenience when DEBUG is True. However, if you need the tasks to be
